requirements.txt
amqp==5.0.6
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
decorator==5.1.0
Django==3.2.7
django-cors-headers==3.10.0
django-mongoengine==0.4.6
django-scopes==1.2.0
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
idna==3.2
ipdb==0.13.9
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
mongoengine==0.23.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.2.0
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.1.0
pymongo==3.12.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
../campuscom-shared-models
../marketplace-shared-lib
../marketplace-shared-models
